hackathon/requirements.txt
streamlit>=1.32.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
faiss-cpu>=1.8.0
pypdf>=4.0.0
huggingface_hub>=0.23.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
requests
